LayoutTests/imported/w3c/web-platform-tests/cookies/meta-blocked-expected.txt

CONSOLE MESSAGE: The Set-Cookie meta tag is obsolete and was ignored. Use the HTTP header Set-Cookie or document.cookie instead.

PASS Cookie is not set from `<meta>`.
